added new controller and service

+++ src/main/kotlin/com/example/customer_order_product/services/OrderService.kt
@@ -1,11 +1,13 @@
 package com.example.customer_order_product.services
 
+import com.example.customer_order_product.dto.OrderDetailDto
 import com.example.customer_order_product.dto.OrderDto
 import com.example.customer_order_product.models.Order
 import com.example.customer_order_product.models.OrderDetail
 import com.example.customer_order_product.models.Product
 import com.example.customer_order_product.repo.OrderRepo
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 
 @Service
 class OrderService(private val orderRepo: OrderRepo) {
@@ -14,32 +16,64 @@ class OrderService(private val orderRepo: OrderRepo) {
        return orderRepo.findByCustomerId(cus_id)
     }
 
+    fun getAllOrders():List<Order>?{
+        return orderRepo.findAll()
+    }
+
+    fun getAllItems(order:Order?):List<OrderDetailDto>?{
+        val orderItemsList = mutableListOf<OrderDetailDto>()
+        order?: return null
+        order.orderDetails?.forEach{
+            orderItemsList
+                .add(
+                    OrderDetailDto(
+                        it.id!!,
+                        it.product.id!!,
+                        it.order.id!!,
+                        it.unitPrice,
+                        it.quantity,
+                        it.total)) }
+        return orderItemsList
+    }
+
     fun removeOrder( order:Order?):Boolean{
         order?: return false
-        orderRepo.delete(order)
-        return true
+        if(!order.payingStatus && !order.paidStatus || order.payingStatus && order.paidStatus){
+            orderRepo.delete(order)
+            return true
+        }
+        return false
     }
 
     fun removeOrderItem( order: Order?, orderDetailId:Long):Boolean{
 
         order?: return false
+
+        if(!order.payingStatus && !order.paidStatus || order.payingStatus && order.paidStatus){
+
         val itemToRemove = order.orderDetails!!.first { it.id == orderDetailId }
-        order.orderDetails!!.remove(itemToRemove)
-        order.totalAmount = order.totalAmount?.minus(itemToRemove.total)
-        if(order.orderDetails?.isEmpty()!!) {
+        order.also {
+            it.orderDetails!!.remove(itemToRemove)
+            it.totalAmount = order.totalAmount?.minus(itemToRemove.total)
+        }
+        if(order.orderDetails?.isEmpty() == true) {
             orderRepo.delete(order)
             return true
         }
         orderRepo.save(order)
         return true
+        }
+        return false
     }
 
     fun addMoreItems(order:Order?, products_quantities:Map<Product?, Int>?):Boolean{
+
         order?: return false
         if(order.payingStatus) return false
         products_quantities?: return false
+
         val orderDetailsList = mutableListOf<OrderDetail>()
-        var total = 0.0
+        var total = BigDecimal.valueOf(0.0)
         val existProduct = order.orderDetails?.map { it.product }
 
         products_quantities.forEach{(product, quantity) ->
@@ -49,13 +83,13 @@ class OrderService(private val orderRepo: OrderRepo) {
                     order.orderDetails?.first { it.product == product }?.also {
                             total -= it.total
                             it.quantity += quantity
-                            it.total = it.quantity * product.unitPrice
-                            total+=it.total
+                            it.total = BigDecimal.valueOf(it.quantity.toLong()) * product.unitPrice
+                            total += it.total
                         }
                 }
                 else{
                     val price = product.unitPrice
-                    val totalPrice = price*quantity
+                    val totalPrice = price*BigDecimal.valueOf(quantity.toLong())
                     total+=totalPrice
                     orderDetailsList.add(OrderDetail(order, product, price, quantity, totalPrice))
                 }
@@ -69,14 +103,17 @@ class OrderService(private val orderRepo: OrderRepo) {
     }
 
     fun updateOrderItems(order: Order?, products_quantities:Map<Product?, Int>?):Boolean{
+
         order?: return false
         if(!order.payingStatus) return false
         products_quantities?: return false
         order.orderDetails?: return false
-        var preTotal = 0.0
-        var newTotal = 0.0
+
+        var preTotal = BigDecimal.valueOf(0.0)
+        var newTotal = BigDecimal.valueOf(0.0)
 
         val productToRemove = mutableSetOf<Product>()
+
         order.orderDetails?.map{ orderDetail ->
             if(orderDetail.product in products_quantities.keys){
                 preTotal+=orderDetail.total
@@ -85,7 +122,7 @@ class OrderService(private val orderRepo: OrderRepo) {
                 }
                 else {
                     orderDetail.quantity = products_quantities[orderDetail.product]!!
-                    orderDetail.total = orderDetail.quantity * orderDetail.product.unitPrice
+                    orderDetail.total = BigDecimal.valueOf(orderDetail.quantity.toLong()) * orderDetail.product.unitPrice
                     newTotal += orderDetail.total
                 }
             }
